+++ src/ethereum/ethereumInfos.ts
@@ -1,3 +1,4 @@
+import { amoy } from './info/amoyInfo'
 import { arbitrum } from './info/arbitrumInfo'
 import { avalanche } from './info/avalancheInfo'
 import { base } from './info/baseInfo'
@@ -11,18 +12,16 @@ import { ethereumpow } from './info/ethereumpowInfo'
 import { fantom } from './info/fantomInfo'
 import { filecoinfevmcalibration } from './info/filecoinFevmCalibrationInfo'
 import { filecoinfevm } from './info/filecoinFevmInfo'
-import { goerli } from './info/goerliInfo'
-import { kovan } from './info/kovanInfo'
-import { mumbai } from './info/mumbaiInfo'
+import { holesky } from './info/holeskyInfo'
 import { optimism } from './info/optimismInfo'
 import { polygon } from './info/polygonInfo'
 import { pulsechain } from './info/pulsechainInfo'
-import { rinkeby } from './info/rinkebyInfo'
-import { ropsten } from './info/ropstenInfo'
 import { rsk } from './info/rskInfo'
+import { sepolia } from './info/sepoliaInfo'
 import { zksync } from './info/zksyncInfo'
 
 export const ethereumPlugins = {
+  amoy,
   arbitrum,
   avalanche,
   base,
@@ -36,14 +35,11 @@ export const ethereumPlugins = {
   fantom,
   filecoinfevm,
   filecoinfevmcalibration,
-  goerli,
-  kovan,
-  mumbai,
+  holesky,
   optimism,
   polygon,
   pulsechain,
-  rinkeby,
-  ropsten,
+  sepolia,
   rsk,
   zksync
 }