+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
+- added: Added Amoy Polygon testnet
+- added: Added Sepolia Ethereum testnet
 - added: Added support for WebSocket connections to Blockbook servers via `BlockbookWsAdapter`
+- remove: Remove deprecated Ethereum testnets: Goerli, Kovan, Rinkeby, and Ropsten.
+- remove: Remove deprecated Mumbai Polygon testnet
 
 ## 4.13.0 (2024-07-16)